Move Game to components dir
